RemoteLogging

Documentation for RemoteLogging.

RemoteLogging.ConsoleRemoteLogger

RemoteLogging.ProgressRemoteLogger

RemoteLogging.RemoteLogger

RemoteLogging.activate_listener

RemoteLogging.console_example_formatter

RemoteLogging.group_module_filter

RemoteLogging.listen_message

RemoteLogging.listen_progress

RemoteLogging.root_module

RemoteLogging.ConsoleRemoteLogger — Method

ConsoleRemoteLogger(; [kwargs...])

Create a logger that sends log messages over TCP as text.

For full help of keyword arguments, see RemoteLogger

To change formatting of the logger, pass a function that accepts IO and returns a FormatLogger using that IO. This formatter will be used as tthe final sink in ConsoleRemoteLogger. Otherwise, Logging.ConsoleLogger is used.

Example

# assuming that a listener has already been set up
logger = ConsoleRemoteLogger(; port=50010)
with_logger(logger) do
    @debug "debug"
    @info "info"
end
close(logger)

In this case, only the info message should be transported over the network.

RemoteLogging.ProgressRemoteLogger — Method

ProgressRemoteLogger(; [host,] [port])

Create a logger that sends ProgressLogging.Progress information over TCP.

Example

# assuming that a listener has already been set up
logger = ProgressRemoteLogger(; port=50011)
with_logger(logger) do
    @progress for i=1:10
        sleep(0.1)
    end
end
close(logger)

The progress message should appear on the remote listener.

RemoteLogging.RemoteLogger — Method

RemoteLogger(; [kwargs...])

A combined ConsoleRemoteLogger and ProgressRemoteLogger. Log messages and progress bars produced with this logger will appear in the remote listener.

Arguments

host: IP address of the listener. The listener should be running in advance.

port: Port of the listener. port and port+1 will be used.

console_displaywidth: intended width of the log viewer. It can be used by the printer. RemoteLogging's default listener does not use this value.

console_loglevel: minimum log level to be sent to the remote console

console_formatter: extra formatter for console. It should be a function that accepts a logger and combine it with a another compositional logger.

console_exclude_group: log groups to ignore. Log messages with this group will not be sent to ConsoleRemoteLogger. Ignores :ProgressLogging by default.

console_exclude_module: source modules to ignore in console. Log messages that used to be hidden because it originates from another library will not be filtered here. Specify them here so that they will not be sent.

RemoteLogging.activate_listener — Function

activate_listener(host, port; [autoclose])

Start both progress and message listener using given host and port, port+1. If block is true, block the console, and clear the screen whenever a new line (enter) is pressed. To close the listener and disconnect the server, use Ctrl+C.

RemoteLogging.console_example_formatter — Method

An exmaple for implementing formatter for ConsoleRemoteLogger.

RemoteLogging.group_module_filter — Method

group_module_filter(logger, ex_group, ex_module)

Combined EarlyFilteredLogger that filters based on module and group of the log.

RemoteLogging.listen_message — Function

listen_message(host, port, [io])

Listen to incoming messages and print them. The formatting of the messages should be set in the RemoteLogger.

RemoteLogging.listen_progress — Method

listen_progress(host, port)

Listen to progress messages and show it.

RemoteLogging.root_module — Method

root_module(m)

Find the root module of the given module. Useful for filtering out modules in log messages.